pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "recursive-llm"
version = "0.1.0"
description = "Recursive Language Models for unbounded context processing"
authors = [{name = "Grigori Gvadzabia"}]
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
keywords = ["llm", "ai", "nlp", "recursive", "language-models"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "RestrictedPython>=6.0",
    "python-dotenv>=1.0.0",
    "dspy-ai>=2.4.0",
    "e2b>=0.17.0",
    "e2b-code-interpreter>=0.0.5",
    "networkx>=3.0",
    "pyvis>=0.3.2",
    "pytest>=8.4.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "black>=24.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/recursive-llm"
Documentation = "https://github.com/yourusername/recursive-llm"
Repository = "https://github.com/yourusername/recursive-llm"
Issues = "https://github.com/yourusername/recursive-llm/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v"

[tool.black]
line-length = 100
target-version = ['py39']

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
